services/graph_harness_api/requirements.txt
fastapi>=0.135,<0.136
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
email-validator>=2.2.0
sqlalchemy>=2.0.0
greenlet>=3.2.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
alembic>=1.13.0
google-cloud-secret-manager>=2.16.0
google-cloud-storage>=2.10.0
httpx>=0.27.0
redis>=5.0.0
artana-kernel @ git+https://github.com/aandresalvarez/artana-kernel.git@main
bcrypt>=4.1.0
cryptography>=46.0.5
defusedxml>=0.7.1
jsonschema>=4.20.0
aiohttp>=3.10.0
anyio>=4.0.0
requests>=2.30.0
PyJWT>=2.8.0
protobuf>=6.33.5
